+++ src/internal/account.rs
@@ -1,7 +1,8 @@
-use crate::internal::{agent::Os, XVCKey, AUTH_HEADER_WITHOUT_AGENT, AUTH_USER_AGENT, LANGUAGE};
+use crate::internal::{XVCKey, AUTH_HEADER_WITHOUT_AGENT, AUTH_USER_AGENT, LANGUAGE};
 use crate::{define, define_host};
 use reqwest::header::{self, HeaderMap, HeaderValue};
 use reqwest::Url;
+use crate::types::Os;
 
 define! {account_path, "account"}
