+++ src/types/mod.rs
@@ -1,12 +1,15 @@
-pub use channel::*;
-pub use chat_room::*;
-pub use user::*;
-
 pub mod channel;
 pub mod chat_room;
 pub mod open_chat;
 pub mod structs;
 pub mod user;
+
+mod os;
+
+pub use channel::*;
+pub use chat_room::*;
+pub use user::*;
+pub use os::*;
 
 pub type LogId = u64;
 pub type Timestamp = u64;
